added @AuthenticationPrincipal to getUser, updateOneUser and deleteUser in UserController

+++ src/main/java/com/marijana/library1223/controllers/UserController.java
@@ -2,10 +2,13 @@ package com.marijana.library1223.controllers;
 
 import com.marijana.library1223.configuration.PaginationConfiguration;
 import com.marijana.library1223.dtos.UserDto;
+import com.marijana.library1223.exceptions.AccessDeniedException;
 import com.marijana.library1223.exceptions.BadRequestException;
 import com.marijana.library1223.services.UserService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
@@ -55,9 +58,17 @@ public class UserController {
 
 
     @GetMapping(value = "/{username}")
-    ResponseEntity<UserDto> getUser(@PathVariable("username") String username) {
-        UserDto optionalUser = userService.getUserByUsername(username);
-        return ResponseEntity.ok().body(optionalUser);
+    ResponseEntity<UserDto> getUser(
+            @PathVariable("username") String username,
+            @AuthenticationPrincipal UserDetails userDetails) {
+
+        if (username.equals(userDetails.getUsername())) {
+            UserDto optionalUser = userService.getUserByUsername(username);
+            return ResponseEntity.ok().body(optionalUser);
+
+        } else {
+            throw new AccessDeniedException("You are not authorized to access this user");
+        }
     }
 
 
@@ -81,18 +92,44 @@ public class UserController {
     }
 
 
-
     @PutMapping("/update/{username}")
-    public ResponseEntity<UserDto> updateOneUser(@PathVariable("username") String username, @RequestBody UserDto userDto) {
-        userService.updateUserPassword(username, userDto);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<UserDto> updateOneUser(
+            @PathVariable("username") String username,
+            @RequestBody UserDto userDto,
+            @AuthenticationPrincipal UserDetails userDetails) {
+
+        if(username.equals(userDetails.getUsername())) {
+
+            userService.updateUserPassword(username, userDto);
+            return ResponseEntity.noContent().build();
+
+        } else {
+
+            throw new AccessDeniedException("You are not authorized to update this user");
+
+        }
+
     }
 
-    @DeleteMapping(value = "/delete/{username}")
-    public ResponseEntity<Object> deleteOneUser(@PathVariable("username") String username) {
-        userService.deleteUser(username);
-        return ResponseEntity.noContent().build();
+
+    @DeleteMapping(value = "/delete/{username}" )
+    public String deleteUser(
+            @PathVariable("username") String username,
+            @AuthenticationPrincipal UserDetails userDetails) {
+
+        if (username.equals(userDetails.getUsername())) {
+
+            userService.deleteUser(username);
+
+            return "User " + userDetails.getUsername() + " has been deleted: " + username;
+
+        } else {
+            throw new AccessDeniedException("You are not authorized to delete this user");
+        }
+
     }
+
+
 
     //-----authorities
